Fix broken build
