+++ psalm.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <psalm
     errorLevel="4"
-    resolveFromConfigFile="true"
     memoizeMethodCallResults="true"
+    errorBaseline="psalm-baseline.xml"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xmlns="https://getpsalm.org/schema/config"
     xsi:schemaLocation="https://getpsalm.org/schema/config vendor/vimeo/psalm/config.xsd"
